fix(release): add perception to package authority
